requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.21.0
torch>=1.9.0
torchvision>=0.10.0 